Move QlFileDes definition to POSIX
